Fixed some shit

+++ include/backtracking.h
@@ -6,7 +6,7 @@
 /*   By: jcaron <jcaron@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/19 18:17:48 by jcaron            #+#    #+#             */
-/*   Updated: 2023/01/21 12:45:31 by jcaron           ###   ########.fr       */
+/*   Updated: 2023/01/26 18:44:22 by jcaron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@
 
 typedef struct s_solution
 {
-	char	*operation;
-	char	*opti;
-	size_t	depth;
-	size_t	max_depth;
+	char			*operation;
+	char			*opti;
+	unsigned long	depth;
+	unsigned long	max_depth;
 }	t_solution;
 
 void	sort_backtrack(t_stack *a, t_stack *b);
